Revert "Use AlertDialog for MediaRouteControllerDialog"
This reverts commit 70c0da386df98e71a9f3487f0b3af5f93b8e0c2a.

Background: To just inherit AlertDialog, the content view should include
a title as we do in support library (AlertDialog uses NO_TITLE feature),
but up-streaming support library implementation to the framework at this
point might cause more issues. Verified that the narrow dialog issue
(b/22044600) does not happen in the framework implementation regardless of
whether it uses AlertDialog or not.

Bug: 22286869
Change-Id: Ic2554cc9e683beff29d1deee91945c1dace83ab1