BatteryStats: Remove uid_cputime Slog.wtf for high cpu times
High cpu times are expected as multiple cores can be running at the
same time, so comparing against the time between samples is incorrect.

I am reasonable certain that the values we see now are correct, so disabling this
check. However, checking for negative values (overflows) is still enabled and
will remain enabled because there is no case where we will be ok with negative deltas.

Bug:22461683
Change-Id: If9c7cdbb75ceaed059d1e0f4dd83cfdd3e021a93