Crash apps with bad notification icons.
This shouldn't happen, since there are many places where
invalid icons should already have been either fixed (in the
case where there's an .icon but no .mSmallIcon) or rejected
(if they're both null or invalid). But if a notification
makes it all the way to SystemUI without a valid icon, let's
crash the sender.

Bug: 23011305
Change-Id: Ifaebec57d59baa1defb4520178b5815d47ed5712