Clear SuggestionsPopupWindow.mSpansLengths after it is used.
This is a no risk optimization that minimizes the time when
strong references to SuggestionSpan and Integer instances are
hold in SuggestionsPopupWindow.mSpansLengths.

SuggestionsPopupWindow.mSpansLengths are used while the given
SuggestionSpans array is being sorted. We don't need to keep
object references there after the sort is completed.

No user visible change is intended with this CL.

Change-Id: I7882d741c08d07ad7f905ad1d171c0d13da39397