Ensure feature flags are set prior to checking Activity Transition.
Bug 23764120

Window features are set in PhoneWindow.installDecor. The Activity
Transition feature is checked after onCreate. If the developer
hasn't accessed the decor View in onCreate, the Activity Transition
feature will not have been set prior to checking it. To correct
this, the decor View will be forced to be created prior to
checking the Activity Transition feature state.

Change-Id: I2c756b6736ae7f3a346f4f9bcef6936371ff03c5