DO NOT MERGE Allow uid to be passed for more operations
This expands get, getmtime, exportKey, getKeyCharacteristcs and begin to
accept a uid to run as. This is only for system to use keys owned by
Wifi and VPN, and not something that can be used to do operations as
another arbitrary application.

Cherry-picked from 250bea9fe6a4513039a96ed226192fc6a2933edd
Bug: 23978113
Change-Id: I06aa089859edc934a5415e3b184b917d6d171ae2