TextViewActions for Espresso-testing TextView.
The plan is to add common TextView actions here so they can be used
in tests.
The locateTextAtIndex method is probably not the most efficient way
to locate a particular spot in the TextView but it works well enough
and doesn't use any private/hidden APIs. This might make it easy to
port this test as CTS.
Ideas on alternative implementations for this method are welcome.

Bug: 24102650
Change-Id: Ifb07cc86c58bd5833c37919fa9294a2374dba2a0