Fix leaking the first activity while preserving windows.
Preserving decor view across activity relaunches would leak the first
activity, because the decor view would hold onto into in the form of the
context. This CL fixes that by having DecorView and NonClientDecorView
use application context instead.

Another source of a leak is DecorView being inner, non static class.
This would keep the orignal, first PhoneWindow around, which in turn
holds a reference to the activity. DecorView is now static and has
explicit reference to the PhoneWindow.

Change-Id: I3df58755d65d3d36ea2157908b0000b2d5c4ab70