Update dock divider display info. whenever it changes.
The docked divider display info. was set on object creation and never
changed. This can hang the dock divider if the device rotation is
changed and the user moves the divider rapidly in and out of the
screen edge closer to the right. We now update the display info.
whenever it changes.

Change-Id: Ie5f0c7ebdafedb156d04eceee00fc7e4176d9bf8