Disassociate system windows from apps
With the original logic, if an app creates a system window, when the
user goes to home screen, the system window will be still there and
become unable to receive input events, because the system window will
be also changed to the stopped state with the app window, and the
current logic of ViewRootImpl forbid a stopped window receiving input
events.

This change prevents assigning the token of the app window to system
windows created by the app, so that when the app goes to the stopped
state, its system windows won't be affected (can still receive input
events).

This change is related to the following changes:
https://android.googlesource.com/platform/frameworks/base/+/286f84c31fb471bb211f0281a7c46d1b48a2bcef
https://android.googlesource.com/platform/frameworks/base/+/13fd279cebfb9b60687f7d98024f93a56e0c50e9
https://android.googlesource.com/platform/frameworks/base/+/3768852f4cce12ef104e3c5ea659202ad76126fb
https://android.googlesource.com/platform/frameworks/base/+/4bb02451e52ec5d8917060758997eb84a3924dc2

Bug: 24967303
Bug: https://code.google.com/p/android/issues/detail?id=189710

Change-Id: I515e47bafcf39a2b1bdf92f11f623bef8fb6263c