SystemUI: fix automatic pairing of BT keyboards
After reset (docking) Pixel C Keyboard that was previously paired with
a device goes into so-called non-discoverable mode, where it will
establish connection only with device that it has connected before. When
scanning for available devices we need to wait till the keyboard starts
advertising itself as discoverable, and only then try to pair.

Also, let's flush the device cache when we attach the base to make sure
the device that we seen before and cached again in the right state after
reset.

Bug: 24915541
Change-Id: I136c1c4235080a25529b4b1c2b1da9bc18508811