Fix recents <-> full screen animations in landscape.
Somewhere along the way we lost the accounting for system decorations
when calculating the crop of the application window during the
recents to/from full screen app transition. This is necessary,
otherwise the animation starts/ends with a black navigation bar and
there is no sense of continuity with the recents thumbnail.

Change-Id: I75bc77cc4badca40b7c59ded39e0d4f72e36c739